remove data confirm 2

+++ client/app/bundles/components/categories/actions/destroy.jsx
@@ -22,19 +22,21 @@ export default class Destroy extends React.Component {
   }
 
   handleDelete(event) {
-    let $target = $(event.target);
-    $target.blur();
-    axios.delete(routes.category_url(this.props.category.id), {
-      params: {
-        authenticity_token: ReactOnRails.authenticityToken()
-      },
-      headers: {'Accept': 'application/json'}
-    })
-    .then(response => {
-      this.props.handleAfterDeleted(this.props.category);
-    })
-    .catch(error => {
-      console.log(error)}
-    );
+    if (confirm(I18n.t('data.confirm_delete'))) {
+      let $target = $(event.target);
+      $target.blur();
+      axios.delete(routes.category_url(this.props.category.id), {
+        params: {
+          authenticity_token: ReactOnRails.authenticityToken()
+        },
+        headers: {'Accept': 'application/json'}
+      })
+      .then(response => {
+        this.props.handleAfterDeleted(this.props.category);
+      })
+      .catch(error => {
+        console.log(error)}
+      );
+    }
   }
 }
